+++ index.ts
@@ -2,5 +2,18 @@ export { default as Parser } from './src/Parser/Parser';
 export { default as SelectBuilder } from './src/SelectBuilder/SelectBuilder';
 export { default as extractGrouped } from './src/extractGrouped/extractGrouped';
 export { default as extractIndexed } from './src/extractIndexed/extractIndexed';
+export { default as extractLookup } from './src/extractLookup/extractLookup';
 export { default as getPagination } from './src/getPagination/getPagination';
+export {
+  runMysql,
+  runMysqlAsync,
+  runMysqlAsyncWithCount,
+  runMysqlWithCount,
+} from './src/runMysql/runMysql';
 export { runPrisma, runPrismaWithCount } from './src/runPrisma/runPrisma';
+export {
+  toSafeJsonRecord,
+  toSafeJsonRecords,
+  toSafeJsonString,
+  toSafeJsonValue,
+} from './src/toSafeJson/toSafeJson';