+++ src/getPagination/getPagination.spec.ts
@@ -17,6 +17,20 @@ describe('getPagination()', () => {
     });
   });
   it('should handle 42 results with offset', () => {
+    const query = new SelectBuilder('SELECT * FROM users LIMIT 10').offset(20);
+    const pagination = getPagination(query, 42n);
+    expect(pagination).toEqual({
+      page: 3,
+      prevPage: 2,
+      nextPage: 4,
+      perPage: 10,
+      numPages: 5,
+      total: 42,
+      isFirst: false,
+      isLast: false,
+    });
+  });
+  it('should handle 42 results with page', () => {
     const query = new SelectBuilder('SELECT * FROM users LIMIT 10').page(3);
     const pagination = getPagination(query, 42n);
     expect(pagination).toEqual({
@@ -34,13 +48,13 @@ describe('getPagination()', () => {
     const query = new SelectBuilder('SELECT * FROM users LIMIT 10');
     const pagination = getPagination(query, 0);
     expect(pagination).toEqual({
-      page: 1,
+      page: null,
       prevPage: null,
       nextPage: null,
       perPage: 10,
       numPages: 0,
       total: 0,
-      isFirst: true,
+      isFirst: false,
       isLast: false,
     });
   });
@@ -48,13 +62,13 @@ describe('getPagination()', () => {
     const query = new SelectBuilder('SELECT * FROM users');
     const pagination = getPagination(query, 42);
     expect(pagination).toEqual({
-      page: 1,
+      page: null,
       prevPage: null,
       nextPage: null,
-      perPage: NaN,
+      perPage: null,
       numPages: 0,
       total: 42,
-      isFirst: true,
+      isFirst: false,
       isLast: false,
     });
   });