+++ src/runPrisma/runPrisma.ts
@@ -1,6 +1,7 @@
 import type SelectBuilder from '../SelectBuilder/SelectBuilder';
 import type { EngineStyle } from '../SelectBuilder/SelectBuilder';
 import getPagination from '../getPagination/getPagination';
+import { toSafeJsonValue } from '../toSafeJson/toSafeJson';
 
 interface PrismaClient {
   $queryRawUnsafe: (
@@ -9,6 +10,13 @@ interface PrismaClient {
   ) => Promise<Array<Record<string, any>>>;
 }
 
+/**
+ * Run the given query with prisma and return the results
+ * @param prisma  The prisma client instance
+ * @param query  The SelectBuilder query to run
+ * @param options
+ * @param options.engine  The engine style to use ("mssql" | "mysql" | "oracle" | "pg")
+ */
 export async function runPrisma(
   prisma: PrismaClient,
   query: SelectBuilder,
@@ -18,6 +26,14 @@ export async function runPrisma(
   return prisma.$queryRawUnsafe(sql, ...bindings);
 }
 
+/**
+ * Run the given query with prisma and return the results, count, and pagination
+ * @param prisma  The prisma client instance
+ * @param query  The SelectBuilder query to run
+ * @param options
+ * @param options.engine  The engine style to use ("mssql" | "mysql" | "oracle" | "pg")
+ * @param options.countExpr  The SQL expression to use for the count (default "*")
+ */
 export async function runPrismaWithCount(
   prisma: PrismaClient,
   query: SelectBuilder,
@@ -31,7 +47,7 @@ export async function runPrismaWithCount(
   if (records.length > 0) {
     const { sql, bindings } = query.compileCount({ engine, countExpr });
     const result = await prisma.$queryRawUnsafe(sql, ...bindings);
-    const total = result[0].found_rows;
+    const total = toSafeJsonValue(result[0].found_rows);
     return { records, total, pagination: getPagination(query, total) };
   } else {
     return { records, total: 0, pagination: getPagination(query, 0) };